实现播放以及进度条，还有小bug

+++ app/src/main/java/com/glandroid/himalaya/interfaces/IPlayerCallback.java
@@ -37,7 +37,7 @@ public interface IPlayerCallback {
      * @param currentProgress
      * @param total
      */
-    void onProgressChange(int currentProgress,long total);
+    void onProgressChange(int currentProgress,int total);
     /**
      * 广告正在加载
      */
